+++ ui-cra/src/components/ResponsiveDrawer.tsx
@@ -14,71 +14,24 @@ import {
   AuthContextProvider,
   coreClient,
   CoreClientContextProvider,
-  OAuthCallback,
   SignIn,
   theme as weaveTheme,
-  V2Routes,
 } from '@weaveworks/weave-gitops';
-import { GitProvider } from '@weaveworks/weave-gitops/ui/lib/api/applications/applications.pb';
-import qs from 'query-string';
 import React from 'react';
-import Lottie from 'react-lottie-player';
-import { Redirect, Route, Switch } from 'react-router-dom';
+import { Route, Switch } from 'react-router-dom';
 import styled from 'styled-components';
 import { Terraform } from '../api/terraform/terraform.pb';
-import error404 from '../assets/img/error404.json';
 import { ReactComponent as MenuIcon } from '../assets/img/menu-burger.svg';
 import { ClustersService } from '../cluster-services/cluster_services.pb';
 import EnterpriseClientProvider from '../contexts/EnterpriseClient/Provider';
 import NotificationsProvider from '../contexts/Notifications/Provider';
 import { TerraformProvider } from '../contexts/Terraform';
-import { Routes } from '../utils/nav';
-import WGApplicationsDashboard from './Applications';
-import AddApplication from './Applications/Add';
-import WGApplicationsBucket from './Applications/Bucket';
-import WGApplicationsFluxRuntime from './Applications/FluxRuntime';
-import WGApplicationsGitRepository from './Applications/GitRepository';
-import WGApplicationsHelmChart from './Applications/HelmChart';
-import WGApplicationsHelmRelease from './Applications/HelmRelease';
-import WGApplicationsHelmRepository from './Applications/HelmRepository';
-import WGApplicationsKustomization from './Applications/Kustomization';
-import WGNotifications from './Applications/Notifications';
-import WGApplicationsOCIRepository from './Applications/OCIRepository';
-import WGApplicationsSources from './Applications/Sources';
-import MCCP from './Clusters';
-import ClusterDashboard from './Clusters/ClusterDashboard';
-import AddClusterWithCredentials from './Clusters/Create';
-import EditCluster from './Clusters/Edit';
+import AppRoutes from '../routes';
 import ErrorBoundary from './ErrorBoundary';
-import { ContentWrapper } from './Layout/ContentWrapper';
-import { PageTemplate } from './Layout/PageTemplate';
 import { Navigation } from './Navigation';
-import Pipelines from './Pipelines';
-import PipelineDetails from './Pipelines/PipelineDetails';
-import Policies from './Policies';
-import PolicyDetails from './Policies/PolicyDetails';
-import PoliciesViolations from './PolicyViolations';
-import PolicyViolationDetails from './PolicyViolations/ViolationDetails';
-import ProgressiveDelivery from './ProgressiveDelivery';
-import CanaryDetails from './ProgressiveDelivery/CanaryDetails';
 import Compose from './ProvidersCompose';
-import TemplatesDashboard from './Templates';
-import TerraformObjectDetail from './Terraform/TerraformObjectDetail';
-import TerraformObjectList from './Terraform/TerraformObjectList';
 
-const GITLAB_OAUTH_CALLBACK = '/oauth/gitlab';
-const POLICIES = '/policies';
-const CANARIES = '/applications/delivery';
-const PIPELINES = '/applications/pipelines';
 export const EDIT_CLUSTER = '/clusters/:clusterName/edit';
-
-function withSearchParams(Cmp: any) {
-  return ({ location: { search }, ...rest }: any) => {
-    const params = qs.parse(search);
-
-    return <Cmp {...rest} {...params} />;
-  };
-}
 
 const drawerWidth = 220;
 
@@ -149,40 +102,6 @@ const SignInWrapper = styled.div`
   }
 `;
 
-const CoreWrapper = styled.div`
-  div[class*='FilterDialog__SlideContainer'] {
-    overflow: hidden;
-  }
-  .MuiFormControl-root {
-    min-width: 0px;
-  }
-  div[class*='ReconciliationGraph'] {
-    svg {
-      min-height: 600px;
-    }
-    .MuiSlider-root.MuiSlider-vertical {
-      height: 200px;
-    }
-  }
-  .MuiButton-root {
-    margin-right: 0;
-  }
-  max-width: calc(100vw - 220px);
-`;
-
-const Page404 = () => (
-  <PageTemplate documentTitle="NotFound" path={[{ label: 'Error' }]}>
-    <ContentWrapper>
-      <Lottie
-        loop
-        animationData={error404}
-        play
-        style={{ width: '100%', height: 650 }}
-      />
-    </ContentWrapper>
-  </PageTemplate>
-);
-
 const App = () => {
   const classes = useStyles();
   const theme = useTheme();
@@ -240,171 +159,7 @@ const App = () => {
         </nav>
         <main className={classes.content}>
           <ErrorBoundary>
-            <Switch>
-              <Route exact path="/">
-                <Redirect to="/clusters" />
-              </Route>
-              <Route component={MCCP} exact path={'/clusters'} />
-              <Route component={MCCP} exact path="/clusters/delete" />
-              <Route
-                component={withSearchParams((props: any) => (
-                  <ClusterDashboard {...props} />
-                ))}
-                path="/cluster"
-              />
-              <Route component={EditCluster} exact path={EDIT_CLUSTER} />
-              <Route
-                component={AddClusterWithCredentials}
-                exact
-                path="/templates/:templateName/create"
-              />
-              <Route
-                component={PoliciesViolations}
-                exact
-                path="/clusters/violations"
-              />
-              <Route
-                component={withSearchParams(PolicyViolationDetails)}
-                exact
-                path="/clusters/violations/details"
-              />
-              <Route
-                component={() => (
-                  <CoreWrapper>
-                    <WGApplicationsDashboard />
-                  </CoreWrapper>
-                )}
-                exact
-                path={V2Routes.Automations}
-              />
-              <Route
-                component={withSearchParams(AddApplication)}
-                exact
-                path="/applications/create"
-              />
-              <Route
-                component={() => (
-                  <CoreWrapper>
-                    <WGApplicationsSources />
-                  </CoreWrapper>
-                )}
-                exact
-                path={V2Routes.Sources}
-              />
-              <Route
-                component={withSearchParams((props: any) => (
-                  <CoreWrapper>
-                    <WGApplicationsKustomization {...props} />
-                  </CoreWrapper>
-                ))}
-                path={V2Routes.Kustomization}
-              />
-              <Route
-                component={withSearchParams((props: any) => (
-                  <CoreWrapper>
-                    <WGApplicationsGitRepository {...props} />
-                  </CoreWrapper>
-                ))}
-                path={V2Routes.GitRepo}
-              />
-              <Route
-                component={withSearchParams((props: any) => (
-                  <CoreWrapper>
-                    <WGApplicationsHelmRepository {...props} />
-                  </CoreWrapper>
-                ))}
-                path={V2Routes.HelmRepo}
-              />
-              <Route
-                component={withSearchParams((props: any) => (
-                  <CoreWrapper>
-                    <WGApplicationsBucket {...props} />
-                  </CoreWrapper>
-                ))}
-                path={V2Routes.Bucket}
-              />
-              <Route
-                component={withSearchParams((props: any) => (
-                  <CoreWrapper>
-                    <WGApplicationsHelmRelease {...props} />
-                  </CoreWrapper>
-                ))}
-                path={V2Routes.HelmRelease}
-              />
-              <Route
-                component={withSearchParams((props: any) => (
-                  <CoreWrapper>
-                    <WGApplicationsHelmChart {...props} />
-                  </CoreWrapper>
-                ))}
-                path={V2Routes.HelmChart}
-              />
-              <Route
-                component={withSearchParams((props: any) => (
-                  <CoreWrapper>
-                    <WGApplicationsOCIRepository {...props} />
-                  </CoreWrapper>
-                ))}
-                path={V2Routes.OCIRepository}
-              />
-              <Route
-                component={() => (
-                  <CoreWrapper>
-                    <WGApplicationsFluxRuntime />
-                  </CoreWrapper>
-                )}
-                path={V2Routes.FluxRuntime}
-              />
-              <Route
-                component={withSearchParams((props: any) => (
-                  <CoreWrapper>
-                    <WGNotifications {...props} />
-                  </CoreWrapper>
-                ))}
-                path={V2Routes.Notifications}
-              />
-              <Route exact path={CANARIES} component={ProgressiveDelivery} />
-              <Route exact path={PIPELINES} component={Pipelines} />
-              <Route
-                exact
-                path="/applications/pipelines/details"
-                component={withSearchParams(PipelineDetails)}
-              />
-              <Route
-                path="/applications/delivery/:id"
-                component={withSearchParams(CanaryDetails)}
-              />
-              <Route exact path={POLICIES} component={Policies} />
-              <Route
-                exact
-                path="/policies/details"
-                component={withSearchParams(PolicyDetails)}
-              />
-              <Route component={TemplatesDashboard} exact path="/templates" />
-              <Route
-                exact
-                path={Routes.TerraformObjects}
-                component={withSearchParams(TerraformObjectList)}
-              />
-              <Route
-                path={Routes.TerraformDetail}
-                component={withSearchParams(TerraformObjectDetail)}
-              />
-              <Route
-                exact
-                path={GITLAB_OAUTH_CALLBACK}
-                component={({ location }: any) => {
-                  const params = qs.parse(location.search);
-                  return (
-                    <OAuthCallback
-                      provider={'GitLab' as GitProvider}
-                      code={params.code as string}
-                    />
-                  );
-                }}
-              />
-              <Route exact render={Page404} />
-            </Switch>
+            <AppRoutes />
           </ErrorBoundary>
         </main>
       </div>
